Use eggtrayicon which has better transparency support (taken from UIM project)
The original eggtrayicon which was bundled with mail-notification ceased
to work properly on modern desktop environments. Transparent icons
didn't work correctly any more. An attempt was done to port
mail-notification to use GtkStatusIcon, but unfortunately this required
a rewrite of a large piece of the mail-notification which isn't feasable
for now. The UIM project has done an excellent job at keeping
eggtrayicon in working condition for both GTK2 and GTK3 environments
so we replace our copy of eggtrayicon with the eggtrayicon code from
the UIM project