Replace gtk_paint_box with cairo calls
The function gtk_paint_box was used to draw a box containing
the number of new mails in the tray icon. For some unknown reason
this has been broken for some time. On modern environments this
function only draws a gray non-transparent box without any borders.
As this makes the tray icon look ugly the code in question was
rewritten using cairo API functions